grocerystore/requirements.txt
asgiref==3.5.2
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
Deprecated==1.2.13
Django==4.1
django-ckeditor==6.4.0
django-js-asset==2.0.0
django-oauth-toolkit==2.1.0
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
oauthlib==3.2.1
packaging==21.3
Pillow==9.2.0
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
